+++ react/components/utils/messages.ts
@@ -141,6 +141,9 @@ export const organizationMessages = defineMessages({
   of: {
     id: `${storePrefix}of`,
   },
+  invalidSchema: {
+    id: `${storePrefix}invalid-schema`,
+  },
   autocompleteSearching: {
     id: `${storePrefix}autocomplete-searching`,
   },