+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added a warning message when the customer schema is invalid
+
 ## [1.11.0] - 2022-06-23
 
 ## [1.10.3] - 2022-06-22